Bug fixes and release controls.

+++ src/screens/EndingScreen.cpp
@@ -37,9 +37,9 @@ namespace screens
 
         renderer.drawRect(mx, my, mw, mh, core::UI::DARKEST_BLUE);
 
-        renderer.drawPlate(mx + mw * core::UI::PLATE_WIDTH_DISP, 
-                          my + mh * core::UI::PLATE_HEIGHT_DISP,   
-                          mw * core::UI::PLATE_WIDTH, 
+        renderer.drawPlate(mx + static_cast<int>(mw * core::UI::PLATE_WIDTH_DISP), 
+                          my + static_cast<int>(mh * core::UI::PLATE_HEIGHT_DISP),   
+                          static_cast<int>(mw * core::UI::PLATE_WIDTH), 
                           core::UI::PLATE_DEFAULT_HEIGHT, 
                           "CONGRATULATIONS!", core::UI::RED);
 
@@ -51,14 +51,14 @@ namespace screens
         };
 
         renderer.drawTextVector(finalMessages, 
-                               mx + mw * core::UI::PLATE_WIDTH_DISP,
-                               my + (mh * core::UI::PLATE_HEIGHT_DISP * 3), 
+                               mx + static_cast<int>(mw * core::UI::PLATE_WIDTH_DISP),
+                               my + static_cast<int>(mh * core::UI::PLATE_HEIGHT_DISP * 3), 
                                core::UI::LIGHT_BLUE, 24);
 
         std::string scoreText = "Final Score: " + std::to_string(m_finalScore);
         renderer.drawText(scoreText,
-                         mx + mw * core::UI::PLATE_WIDTH_DISP,
-                         my + (mh * core::UI::PLATE_HEIGHT_DISP * 4),
+                         mx + static_cast<int>(mw * core::UI::PLATE_WIDTH_DISP),
+                         my + static_cast<int>(mh * core::UI::PLATE_HEIGHT_DISP * 4),
                          core::UI::WHITE_COLOR);
     }
